Cria módulo para depósito
