+++ src/sistema_bancario/__init__.py
@@ -1,27 +1,18 @@
 from typing import NewType, Any
 from . import exceptions
+from . import operacao
+from . import deposito
 
-Operacao = NewType("Operacao", dict[str, Any])
-Deposito = NewType("Deposito", Operacao)
-Saque = NewType("Saque", Operacao)
+Saque = NewType("Saque", operacao.Operacao)
 Conta = NewType("Conta", dict[str, Any])
 
-def criar_deposito(valor: float) -> Deposito:
-    if valor <= 0:
-        raise exceptions.DepositoInvalidoException("Depósito deve ter valor positivo")
-    return Deposito(Operacao({"tipo": "Depósito", "valor": valor}))
 
 def criar_saque(valor: float) -> Saque:
-    return Saque(Operacao({"tipo": "Saque", "valor": valor}))
+    return Saque(operacao.Operacao({"tipo": "Saque", "valor": valor}))
 
 def valor_saque(saque: Saque) -> float:
     return saque["valor"]
 
-def tipo_operacao(operacao: Operacao) -> str:
-    return operacao["tipo"]
-
-def valor_deposito(deposito: Deposito) -> float:
-    return deposito["valor"]
 
 def criar_conta() -> Conta:
     return Conta({"saldo": 0.0,
@@ -45,9 +36,9 @@ def quantidade_saques_do_dia(conta: Conta) -> int:
     return conta["quantidade_saques_do_dia"]
 
 
-def realizar_deposito(conta: Conta, deposito: Deposito) -> None:
-    conta["operacoes"].append(deposito)
-    conta["saldo"] += valor_deposito(deposito)
+def realizar_deposito(conta: Conta, depo: deposito.Deposito) -> None:
+    conta["operacoes"].append(depo)
+    conta["saldo"] += deposito.valor_deposito(depo)
 
 
 def valor_maximo_saque(conta: Conta) -> float:
