+++ tests/test_saque.py
@@ -1,6 +1,6 @@
-from src.sistema_bancario import criar_saque, tipo_operacao, valor_saque
+import src.sistema_bancario as app
 
 def test_deve_extrair_dados_do_saque() -> None:
-    saque = criar_saque(100.00)
-    assert tipo_operacao(saque) == "Saque"
-    assert valor_saque(saque) == 100.00
+    saque = app.criar_saque(100.00)
+    assert app.operacao.tipo_operacao(saque) == "Saque"
+    assert app.valor_saque(saque) == 100.00
